refactor: 거리계산 클래스 추출
