+++ backend/src/main/java/com/NBE4_5_SukChanHoSu/BE/domain/likes/service/UserLikeService.java
@@ -7,6 +7,7 @@ import com.NBE4_5_SukChanHoSu.BE.domain.likes.entity.Matching;
 import com.NBE4_5_SukChanHoSu.BE.domain.likes.entity.UserLikes;
 import com.NBE4_5_SukChanHoSu.BE.domain.likes.repository.MatchingRepository;
 import com.NBE4_5_SukChanHoSu.BE.domain.likes.repository.UserLikesRepository;
+import com.NBE4_5_SukChanHoSu.BE.domain.recommend.service.CalculateDistance;
 import com.NBE4_5_SukChanHoSu.BE.domain.recommend.service.RecommendService;
 import com.NBE4_5_SukChanHoSu.BE.domain.user.dto.response.UserProfileResponse;
 import com.NBE4_5_SukChanHoSu.BE.domain.user.entity.Gender;
@@ -43,6 +44,7 @@ public class UserLikeService {
     private final RedisTTL ttl;
     private final RecommendService matchingService;
     private final ObjectMapper objectMapper;
+    private final CalculateDistance calculateDistance;
 
     private static final String LIKE_STREAM = "like";
     private static final String MATCHING_STREAM = "matching";
@@ -138,7 +140,7 @@ public class UserLikeService {
         cancelLikes(toUser, fromUser);
 
         // 응답 생성
-        int distance = matchingService.calDistance(fromUser, toUser);
+        int distance = calculateDistance.calDistance(fromUser, toUser);
         return new MatchingResponse(matching, distance);
     }
 
@@ -157,7 +159,7 @@ public class UserLikeService {
                     Map<String, Object> map = (Map<String, Object>) value;  // Map(키-값 쌍)으로 캐스팅
                     try {
                         UserLikes like = objectMapper.convertValue(map, UserLikes.class); // Map -> UserLikes 클래스로 변환
-                        int radius = matchingService.calDistance(like.getFromUser(), like.getToUser());  // 거리 계산
+                        int radius = calculateDistance.calDistance(like.getFromUser(), like.getToUser());  // 거리 계산
                         likesUsers.add(new UserProfileResponse(like.getToUser(), radius, like.getCreatedAt()));  // 메모리에 추가
                     } catch (IllegalArgumentException e) {
                         throw new RedisSerializationException("500", "JSON 역직렬화 실패");
@@ -169,7 +171,7 @@ public class UserLikeService {
                 UserProfile likedUser = like.getToUser();   // 좋아요를 받은 사용자
                 if (likedUser != null) {
                     // 내가 좋아요 한 사용자 리스트에 추가
-                    int radius = matchingService.calDistance(like.getFromUser(), like.getToUser());  // 거리 계산
+                    int radius = calculateDistance.calDistance(like.getFromUser(), like.getToUser());  // 거리 계산
                     likesUsers.add(new UserProfileResponse(likedUser, radius, like.getCreatedAt()));  // 메모리에 추가
                     // 캐싱
                     String key = "likes:" + user.getUserId() + ":" + likedUser.getUserId();
@@ -199,7 +201,7 @@ public class UserLikeService {
                     Map<String, Object> map = (Map<String, Object>) value;
                     try {
                         UserLikes like = objectMapper.convertValue(map, UserLikes.class);
-                        int distance = matchingService.calDistance(like.getFromUser(), like.getToUser());    // 거리 계산
+                        int distance = calculateDistance.calDistance(like.getFromUser(), like.getToUser());    // 거리 계산
                         likedUsers.add(new UserProfileResponse(like.getFromUser(), distance,like.getCreatedAt()));
                     } catch (IllegalArgumentException e) {
                         throw new RedisSerializationException("500", "JSON 역직렬화 실패");
@@ -211,7 +213,7 @@ public class UserLikeService {
                 UserProfile likesUser = like.getFromUser(); // 좋아요를 보낸 사용자
                 if (likesUser != null) {
                     // 나를 좋아요한 사용자 리스트에 추가
-                    int distance = matchingService.calDistance(user, likesUser);    // 거리 계산
+                    int distance = calculateDistance.calDistance(user, likesUser);    // 거리 계산
                     likedUsers.add(new UserProfileResponse(likesUser, distance,like.getCreatedAt()));
 
                     // 캐싱
@@ -246,7 +248,7 @@ public class UserLikeService {
                         try {
                             Matching matching = objectMapper.convertValue(map, Matching.class);
                             System.out.println("추출한 유저: " + matching.getFemaleUser());
-                            int distance = matchingService.calDistance(user, matching.getFemaleUser());
+                            int distance = calculateDistance.calDistance(user, matching.getFemaleUser());
                             responses.add(new UserProfileResponse(matching.getFemaleUser(), distance, matching.getCreatedAt()));
                         } catch (IllegalArgumentException e) {
                             throw new RedisSerializationException("500", "JSON 역직렬화 실패"+e.getMessage());
@@ -258,7 +260,7 @@ public class UserLikeService {
                 List<Matching> matches = matchingRepository.findByMaleUser(user);
                 for (Matching matching : matches) {
                     // 매칭된 여자 유저 리스트에 등록
-                    int distance = matchingService.calDistance(user, matching.getFemaleUser());
+                    int distance = calculateDistance.calDistance(user, matching.getFemaleUser());
                     responses.add(new UserProfileResponse(matching.getFemaleUser(), distance, matching.getCreatedAt()));
 
                     // 캐싱
@@ -280,7 +282,7 @@ public class UserLikeService {
                         Map<String, Object> map = (Map<String, Object>) value;
                         try {
                             Matching matching = objectMapper.convertValue(map, Matching.class);
-                            int distance = matchingService.calDistance(user, matching.getMaleUser());
+                            int distance = calculateDistance.calDistance(user, matching.getMaleUser());
                             responses.add(new UserProfileResponse(matching.getMaleUser(), distance, matching.getCreatedAt()));
                             System.out.println("추출한 유저: " + matching.getMaleUser());
                         } catch (IllegalArgumentException e) {
@@ -293,7 +295,7 @@ public class UserLikeService {
                 List<Matching> matches = matchingRepository.findByFemaleUser(user);
                 for (Matching matching : matches) {
                     // 매칭된 남자 유저 리스트에 등록
-                    int distance = matchingService.calDistance(user, matching.getMaleUser());
+                    int distance = calculateDistance.calDistance(user, matching.getMaleUser());
                     responses.add(new UserProfileResponse(matching.getMaleUser(), distance, matching.getCreatedAt()));
 
                     // 캐싱
