+++ backend/src/main/java/com/NBE4_5_SukChanHoSu/BE/domain/recommend/service/CalculateDistance.java
@@ -1,0 +1,36 @@
+package com.NBE4_5_SukChanHoSu.BE.domain.recommend.service;
+
+import com.NBE4_5_SukChanHoSu.BE.domain.user.entity.UserProfile;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CalculateDistance {
+
+    static final double EARTH_RADIUS = 6371; // 지구의 반지름 (단위: km)
+
+    // 거리 계산
+    public int calDistance(UserProfile userProfile1, UserProfile userProfile2) {
+        // 내 위/경도
+        double lat1 = userProfile1.getLatitude();
+        double lon1 = userProfile1.getLongitude();
+        // 탐색 대상의 위/경도
+        double lat2 = userProfile2.getLatitude();
+        double lon2 = userProfile2.getLongitude();
+
+        // 위/경도 차이
+        double dLat = Math.toRadians(lat2 - lat1);
+        double dLon = Math.toRadians(lon2 - lon1);
+
+        // 곡률 감안해서 어쩌구저쩌구해서 거리 계산
+        double a = Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+                Math.cos(Math.toRadians(lat1)) * Math.cos(Math.toRadians(lat2)) *
+                        Math.sin(dLon / 2) * Math.sin(dLon / 2);
+        double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+
+        // 거리를 km로 계산
+        double distance = EARTH_RADIUS * c;
+
+        // 소수 첫째 자리에서 반올림 후 int형으로 반환
+        return (int) Math.round(distance); // 반올림하여 int형으로 반환
+    }
+}
